# feat(gatsby): Sanitize redux page state