@tyhopp 2022-07-07
> Any noticeable changes on performance?

Added a TLDR to the benchmarks section in the description.

There is a marginal performance improvement but not sure how much I trust that given the small sample size. At any rate it doesn't appear to be slower or more memory intensive with the change